Refactored a lot of code to use more pointers and hopefully simplify the code base. I tried to incorporate a system that would make integrating between tiles and structures / resources easier. Not entirely sure if it will work out yet though. EntityPanel allows for navigation, though I have yet to refactor the code. The current schematic might need a looking into because I'm not sure the exact procedure Command Buttons should be able to interact with the game. Also, we need to figure out how to remove a majority of these switch case statements because it will blow out of proportion if we don't tackle it soon.

+++ src/helper.cpp
@@ -53,7 +53,7 @@ int *readBMP(char *filename)
     return tile_info;
 }
 
-VertexArray generateVertices(int xposition, int yposition, int xsize, int ysize, Texture texture) {
+VertexArray generateVertices(float xposition, float yposition, float xsize, float ysize, Texture texture) {
     VertexArray vertices = VertexArray(Quads, 4);
 
     /** Corner positions for the rendering box */
